Cargo.toml: Pin accurate versions in dependencies specification
The version spec for dependencies was sloppy (it failed to specify minor
and patch levels).

Keep in mind that in cargo, dependency of type:

    foo = "x.y.z"

means that "foo" must resolve with version x'.y'.z' where x' == x,
y' >= y, and z' >= z; and that the default dependency resolver will
maximize y' and z'. That is, cargo assumes that projects are following
semver and that you want the latest package which meets your desired
API.

If you really needed a specific version you could say:

    foo = "=x.y.z"

which constrains the resolver.

[ Quentin: Edit commit object and description ]

Signed-off-by: Quentin Monnet <qmo@qmon.net>