CI: Always run doc workflow, but skip build/publish when relevant
Run the documentation workflow unconditionally.

This is so we can enforce successful status check as a requirement for
merging Pull Requests: this feature requires workflows to actually be
_successful_ and not _skipped_.

Here we use a dedicated action to check for the modified paths for the
Pull Request (or commits, when running after a push). If no relevant
paths were updated, we skip the build and deployment. If documentation
was changed, then we do build and deploy.

Signed-off-by: Quentin Monnet <qmo@qmon.net>