ci: Add summary step to dev.yml and sterile.yml workflows
We want to mark both workflows as required-to-pass for merging Pull
Requests. The way GitHub offers this is to reference jobs individually,
by their name. One issue is when the job names contain "dynamic" values
generated from workflow matrix entries, such as a version number for a
given component.

This commit adds summary steps to dev.yml and sterile.yml. This lets us
hook the requirements to the summary step of each workflow, as opposed
to dynamically named steps.

Renaming jobs is an alternative solution that would probably work, but
we prefer having the summaries for now.

Addresses: https://github.com/githedgehog/dataplane/issues/70

[ Quentin: Edit commit object and description ]

Signed-off-by: Quentin Monnet <qmo@qmon.net>