ci: Rebuild docs on manual dispatch
The documentation workflow is supposed to run on Pull Requests only if
the PR touches files relevant to the documentation. This is great, but
the check gets in the way if we trigger the workflow manually.

Ensure that the docs are always built when we trigger the documentation
workflow with a manual dispatch.

[ Quentin: Edit commit object and description ]

Signed-off-by: Quentin Monnet <qmo@qmon.net>