dpdk/queue: Avoid cloning config object
In TX queue's "configure()", we do not otherwise consume the "config"
object that we get as an argument. We can simply reuse it to craft the
return TxQueue object rather than cloning the struct - just like we do
for RxQueue.

Found in Clippy's reports (Clippy suggested to pass "config" by
reference).

Signed-off-by: Quentin Monnet <qmo@qmon.net>