ci: Fix CI permissions causing main dev build to fail
In the dev.yml workflow, the main job, "build", runs only if a Pull
Request changes files that are relevant to the build, but not if the
developer touches only, say, to documentation sources.

The check to determine whether the build should run is based on a
specific job that looks at the modified files. We had an issue: when the
workflow runs on a "push" event, this job fails to analyse the list of
updated files properly. This causes the job to fail, and the "build" job
to be always skipped.

The reason for that is that the job misses the "contents: read"
permission. The permission is usually on by default, but by adding a
block to the job definition we reset the list to the provided values,
only ("pull-requests: read"). We need to give the job permission to read
the contents, too.

[ Quentin: Edited commit object and description ]

Fixes: 0ab2d9c7a319 ("ci: Skip dev.yml, sterile.yml if there are no code changes")
Signed-off-by: Quentin Monnet <qmo@qmon.net>