Revert "ci: Remove summary steps"
This reverts commit 6d69a4e55096fd1034d361d6f584c098b5853dea.

I thought we could have the individual jobs marked as required. But we
hit the following issue:

Individual jobs (build, test, push in dev.yml and sterile.yml) use a
matrix to run with multiple Rust toolchains (stable/nightly). The Rust
toolchain appears in the job name, whether we specify it or not. Then
when trying to fill the list of required checks in GitHub's interface,
we've got to use these names and as far as I can tell this must be an
exact match, so we need to specify the toolchain. So far this looks good
because ideally we'd like to have the checks using the stable toolchain,
not nightly, as required. However, when the job is skipped, because no
relevant files are touched in the Pull Request, the matrix is not
generated, and the toolchain name is not appended to the job name as one
might expect. In such a case, it's no longer possible to match the
required checks with the titles of the status checks for the Pull
Request, and the Pull Request is blocked. Let's re-add the summaries so
we can mark them as required.

Signed-off-by: Quentin Monnet <qmo@qmon.net>