justfile: Hide DOCKER_SOCKET command output
The justfile would use a command snippet starting with "set -x" to
determine the path of the Docker socket: this helps for debugging, but
most of the time it only prints cryptic messages to the console each
time we run just.

Make just print the output only when we set "debug=true". We cannot use
a variable in the triple-backquoted string, so we need a level of
indirection with a separate command and a call to shell(). We also need
to append a semicolon to "set -x" when set, and we use replace_regex()
for this.

Signed-off-by: Quentin Monnet <qmo@qmon.net>