Adjust QA rules for upcoming parse code
The big change is that `forbid(unsafe_code)` and `no_std` are not viable modes after the upcomming expansion.

The primary reason for this is that we will make significant use of const-evaluated unsafe code (which is benign) but which will offend the linter.

no_std was always just a crude means of forbidding IO and discouraging memory allocations in the net crate.
Given that the etherparse error handling requires std, this is no longer useful.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>