Remove academic tooling from design docs
When I created the design-docs build I grabbed the version of my js/css which supported various academic tools such as bibliography generation, paragraph level citation, equation / figure numbering.

None of that is particularly useful in this context, and it adds a fair amount of complexity to an otherwise innocent mdbook setup.

The fix is easy: substitute in the non-academic version of my scripts.

I don't see any important difference in our renders from this change.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>