routing: add interface submodule
This module models the state used to represent interfaces in
some interface table (Iftable).

Interfaces:
  - are identified by an ifindex (u32).
  - have a mandatory name, and optional description
  - have operational and administrative state, that should mirror
    the actual state of real interfaces (e.g. kernel devices)
  - have type-specific information (e.g. vlan tag in 802.1q)
  - can have an arbitrary number of IPv4 and IPv6 addresses.
  - can be attached / detached to / from a VRF.

Interface attachment is an exclusive membership property:
an interface may be attached to only one VRF. In this commit,
attachment is represented as an optional Arc<>  to a dummy VRF
object type (to be re-defined). Once the VRFs are defined,
the Option<> wrapping the Arc<VRF> may be removed if we always
assume the presence of a default VRF to which interfaces are
implicitly attached by default.

Concurrency: not addressed yet. The IfTable may be wrapped
with a RwLock, since it may be modified by some mgmt interface
(to create interfaces), or the interface manager, and by the RPC
with the control-plane, to add/remove addresses.
The Arc<Vrf> is meant to avoid the need to perform a lookup
in order to access the VRF that an interface is attached to.
However, the access to Interfaces (other than from the IfTable)
is not yet addressed.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>