Adjust "inner" naming convention to tuple struct convention
Feedback is that `inner` is a confusing name for these fields.
`inner` was chosen on the grounds that there is an outer and an inner type (i.e. a transparent wrapper type which qualifies the wrapped inner type).

Still, if `.0` is easier for folks to think about then I won't argue.

This should have been done as part of commit 695633370144fd1e9d766974808217cf0a4b8942 (see https://github.com/githedgehog/dataplane/commit/695633370144fd1e9d766974808217cf0a4b8942) but I did not think about it at the time

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

# Conflicts:
#	net/src/ipv6/mod.rs