routing: add Adjacency type and Adjacency table
Add minimal definitions for adjacency and adjacency table. We'll
use the adjacency table to map directly connected next-hops to
MAC addresses for the L2 dst MAC re-write. This table should play
the role of an ARP/neigh cache, and we may populate it from the
kernel neighbor system, or via netlink.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>