Make nightly optional
It seems a little extreme to assume that we will always pass every build and lint from nightly.

That said, nightly serves as a great early warning system.
It is always nice to know if something is likely going to be deprecated, and the nightly lints are typically good ideas even if they aren't "stable" yet.

Thus, mark nightly failures as "yellow flags"

With this adjustment to dev.yml we should get an annotation, but not a CI failure, in the event that nightly fails.

`beta` and `stable` build / test / doc failures are still blocking for release.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>