feat(cli): separate local from remote args
In previous commits, RequestArgs included all args, whether the
action they applied to was local or remote. This meant that all
args (e.g. connect path) were included and sent in CliRequest
to the remote. This patch defines CliArgs as the main struct were
arguments are stored, which embeds a RequestArgs (for args for
remote actions), which are the only ones that get sent to the
remote.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>