feat(net): Introduce `TestBuffer`
Add an optional feature called `test_buffer` which, if enabled, includes a `TestBuffer` struct in `net`.

`TestBuffer` implements `PacketBuffer`/`PacketBufferMut` traits as well as the `Tailroom` traits.

This makes it suitable as a standin for `Mbuf` in dataplane tests.

The key advantage is that you can create a `TestBuffer` without setting up a whole EAL.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>