build!: Swap to 2024 edition
Jump to the 2024 edition of rust.

This commit sets our minimum supported rust version (MSRV) to 1.85.0, and will break builds on machines which have yet to run `rustup update`.

Note that gen is a reserved keyword in 2024 so we need to use the (unfortunate) raw ident trick to continue using bolero (thus all the `r#gen` changes).

The change to dpdk-sys.env pulls in the dpdk-sys change which includes rust 1.85.0.

Developers will need to run `just refresh-compile-env` to continue to build the project.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>