fix(dependabot): dependabot should also suggest major versions
It is likely that many of the generated PRs will need to be rejected and redone done by hand (given that they are breaking by definition).

That said, this functions as an automatic detection mechanism for dependencies which do a major version bump.
We have no special reason to sleep on those changes; they aren't going to get any easier if we keep developing against deprecated or removed APIs.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>