bump(dpdk-sys): switch to non-musl version
We recently did a largeish rework of dpdk-sys.
It seems to work well for me based on current local testing but we should confirm in CI and on other dev machines before we move on.

If the commits of [dpdk-sys #92](https://github.com/githedgehog/dpdk-sys/pull/92) need to be reverted then it isn't going to get easier by waiting.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>