ci: List acceptable commit types in commitlint's workflow file
I never remember what commit subject types the linter allows, and it's a
pain to search for them in the browser. Let's list them directly in the
workflow file, where everyone can easily find them.

Mid-term, this list should go into a contributing guide, but we don't
have that yet.

Signed-off-by: Quentin Monnet <qmo@qmon.net>