feat(routing/prefix): Add Prefix.covers(), .covers_addr() methods
Add two new methods to the Prefix struct to tell whether a given prefix
"covers" an IP address, or another prefix, respectively.

A prefix "covers" an IP address if this address is part of the range
defined by the prefix. It "covers" another prefix if all addresses in
the latter also belong to the former, in other words, if the second
prefix is a subset of the former.

These methods are meant to be used for NAT in subsequent commits.

Signed-off-by: Quentin Monnet <qmo@qmon.net>