refactor(nat): Handle VPCs by VNI rather than by name
For now we assume a 1:1 mapping between VNIs and VPCs, and we use VNIs
(integers) rather than strings to reference these VPCs. It's simpler,
more efficient, and makes more sense: When a packet arrives, we know its
source VNI, not the associated VPC name before we lookup based on the
VNI. As it turns out, the only lookup we'll need for the early NAT
implementation will be based on VNI anyway.

Signed-off-by: Quentin Monnet <qmo@qmon.net>