feat(nat): Add Nat pipeline network function
Now that we have a Nat object with all the logics to do the address
translation, implement the NetworkFunction trait so we can use it as
part of dataplane pipelines.

Note that at this stage, the new Nat network function is subject to the
limitations of the Nat module: in particular, it is not functional yet,
because the VNI lookup has not been implemented.

Signed-off-by: Quentin Monnet <qmo@qmon.net>