feat(nat/fabric): Implement peering policies representation
Add a struct PeeringPolicy to represent the link between two PIFs.
This object is meant to be used for NAT in a subsequent commit. Knowing
the source VNI and the destination PIF for a packet, we can use peering
policies to retrieve the source PIF for the packet:

- Source VNI is obtained by a lookup in the VXLAN header
- Source VNI provides source VPC
- Destination IP provides destination PIF
- There is at most one peering policy linking this destination PIF to a
  source PIF on the source VPC containing the endpoint prefix covering
  the source IP for the packet

Currently, the "API" for adding or looking up peering policies is not
refined; it's expected that both the interface to represent peering
policies and its exposed methods will move a lot.

Signed-off-by: Quentin Monnet <qmo@qmon.net>