chore: Commit Cargo.lock changes
Building the repo (or running Clippy on it!) produces changes to the
Cargo.lock, that should likely have been committed in a recent PR. This
is annoying locally, and in CI where it may break the per-commit build
(with the script refusing to loop on commits for a dirty branch). Let's
commit the missing update.

Signed-off-by: Quentin Monnet <qmo@qmon.net>