fix(nat): Fix address mapping for translation
In the previous episode... IpList was an interface that would provide a
flattened representation of all IP addresses from a set of prefixes, and
we'd use that to determine a current and a target lists of IP addresses,
so we could do the translation from the former towards the latter. The
only constraint (apart from IP versions) was that both resulting lists
should have the same size.

This design is broken; it doesn't account for network and broadcast
addresses when flattening the various prefixes, meaning we can end up
mapping a valid unicast address to subnet broadcast address, for
example, by accident. See the documentation that is being removed in
that commit, in particular the second diagram with the case where
10.2.0.1 maps to 10.8.1.0, the latter potentially being a network
address.

To address this issue, we must tighten the constraints: instead of
having IpList represent the whole flattened set of prefixes, it now
represent just one prefix, which is (for the initial list) the prefix
the IP address to translate belongs to; and for the target list, the
corresponding prefix with an identical size, meaning that the start and
end address will be the same and we won't risk mapping to network or
broadcast addresses when picking IP addresses in the middle of the
range.

Concretely, the structure and semantics of the IpList change completely
(single prefix instead of flattened representation for multiple
prefixes). Notably, the way we create the IpList objects changes: we
drop IpList::from_prefixes() in favour of IpList::generate_ranges(),
which must also be aware of the destination IP given that it must find
the prefix containing it.

This commit is an intermediate state before further changes, so we
remove the docs and unit tests for the iplist submodule but don't
replace them just yet.

Signed-off-by: Quentin Monnet <qmo@qmon.net>