feat(nat): Extend IpList to account for excluded prefixes
Initially, the IpList struct represented a flattened set of prefixes. In
the previous commit, we completely restructured the submodule to make
the object represent a single prefix instead. This commit brings
additional changes.

This commit is based on the current version of the proposed API for the
Gateway [0]. In particular, let's consider the following example from
the document:

    apiVersion: gateway.githedgehog.com/v1alpha1
    kind: Peering
    metadata:
      name: vpc-1--vpc-2
    spec:
      peering:
        vpc-1:
          ips:
            - cidr: 10.1.1.0/24
            - not: 10.1.1.42/32 # that's just a syntactic sugar to avoid enumerating all the subnets explicitly
          as:
            - 192.168.1.0/24
            - not: 192.168.1.7/32 # that's just a syntactic sugar to avoid enumerating all the subnets explicitly
        vpc-2:
          ips:
            - cidr: 10.1.1.0/24
          as:
            - 192.168.2.0/24

We note that the list of endpoint prefixes ("ips") and corresponding
address sets for NAT ("as") can be expressed as prefixes ("cidrs") to
which we can "exclusion prefixes" ("not").

This commit is a first attempt at supporting this semantics in the nat
module, and in particular in the IpList submodule, by accounting for
exclusion prefixes when retrieving addresses at a given offset, or the
offset of a given address from an IpList. Note that the code is not
exercised yet, because we don't have a configuration step that accounts
for these exclusion prefixes, but we'll provide unit tests in a
follow-up commit.

[0] https://github.com/githedgehog/gateway/blob/70010a597b325fb3a3948c73dce44c5865fc973a/docs/proposed-api.md

Signed-off-by: Quentin Monnet <qmo@qmon.net>