test(nat): Add new unit tests for IpList
We recently reworked the iplist submodule and changed the representation
provided by the object, making all the unit tests moot; they were
removed in a previous commit. This commit adds a new set of unit tests
for the submodule. They did help fix a few issues in the code before
opening the Pull Request.

Signed-off-by: Quentin Monnet <qmo@qmon.net>