build(justfile): Warn user when compile-env image is missing/outdated
The idea behind this check is to warn users by printing to stderr from
the justfile when their compilation environment is outdated. It works,
but with a few (minor) caveats:

- It warns users even on the first run, when no compile-env at all is
  present, even though it's sort of expected that the
  compile-environment is missing in that case.

- It warns users even on invocations of "just refresh-compile-env". I
  have not found a way to access to the recipe name that the user used
  to call just, and to avoid printing the message in that case.

- Similarly, it warn the user when invoking other commands that might
  not, in fact, require the latest compilation environment to be
  present.

The hope is that even with the following limitations, it remains a net
gain to tell developers when their environment might be outdated.

Signed-off-by: Quentin Monnet <qmo@qmon.net>