fix: make eth optional
It turns out that (upcoming) vxlan encap / decap benefit from making eth an optional field.
Specifically, at one point in the pipeline we simply don't know the ethernet header we want to use yet.

It is also entirely possible that we will use layer 3 tunneling for other protocols in the future so making eth required is not generally wise.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>