fix(net/ethtype): Make serde attributes dependent on "serde" feature
The use of dependency crate "serde" in the "net" package is optional,
based on the "serde" feature for the package. As a consequence, we
cannot always derive the serde traits without taking precautions inside
of the package.

This issue has crept in in spite of the CI checks, because CI does not
build every package independently. We observed the build failure with:

    $ just cargo build -p net

We then found the relevant commit that introduced the issue by
bissecting, and observed we could fix the issue, on top of that commit,
by enabling "serde" by running instead:

    $ just cargo build -p net -F serde

However, the issue doesn't appear at all by running only:

    $ just cargo build

This is likely because when building from the root of the repository, we
must address requirements from all packages, including the "dataplane"
package that requires the "net" package to be built with its "serde"
feature, thus masking the issue.

Anyway, fix the bug by making traits derivation conditional to the use
of the "serde" feature in ethtype.rs.

Fixes: 55d2b2fc46b5 (feat: convenience methods for ethtype)
Reported-by: Fredi Raspall <fredi@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>