feat: introduce test-runner.sh
Cargo automatically runs this script for every unit test (this applies to nextest as well).
The script has two main responsibilities:

It runs setcap on the test binary to elevate the test's permitted capabilities.
This action does not cause the tests to run with these capabilities active by default.
That would involve setting the effective capabilities for the test binary (which we don't do).
Instead, assigning the permitted capabilities allows the use of the caps crate to allow us to request elevated
permissions for specific sections of test code.

The purpose of these elevated privileges is to allow the tests to create and destroy virtual network interfaces and
network namespaces (as is required for integration testing).

It bind mounts the (setcap modified) test binary, the project directory, and a few other files into a (read-only)
docker container (which executes the test). This docker container contains only libc and libgcc_s (to better
simulate our deployment environment and discourage faulty assumptions about what will be available at runtime).

The purpose of this container is to

minimize the damage a faulty test might do
make sure that we aren't relying on resources only available on the developer's machine in the tests (test like
we are in prod).
Hopefully, this process also requires us to carefully think about what parts of our code require which privileges (and
to document these requirements carefully).

P.S.

test-runner.sh starts the tests in a different working directory, so we need to adjust the hash.rs tests slightly to account for that.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>