fix: ahash dependencies
ahash is a bit out of date on zerocopy and it is causing duplicate dependencies to crop up.

I submitted a patch here: https://github.com/tkaitchuck/aHash/pull/260

but it sounds like it may be a bit till that is resolved.

until then, we should just point at the fork.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>