ci: Remove "continue-on-error" for builds with Nightly toolchain
The builds with the Nightly toolchain are no longer marked as "required"
for the Pull Request to be merged, or the merge queue to go through. We
can make the job fail when one step fails. Even though we rely on the
last "Summary" step, the job with Nightly will be skipped when building
the summary (because it's still marked as "optional"), failing the job
shouldn't be blocking for PR validation/merge.

The objective of this commit is to get feedback on the failure from the
list of status checks at the end of the Pull Request page on GitHub,
rather than intrusive comments in the middle of the discussions (next
commit will remove the creation of such comments).

Signed-off-by: Quentin Monnet <qmo@qmon.net>