refactor(routing): Rename package as "dataplane-routing"
We're trying to set up the FOSSA tool to get license and security checks
for the repository, but the tools is being silly, and raises an issue on
the "routing" crate: but the incompatible license is with the "routing"
crate from from https://crates.io, completely different from ours, and
not our own "routing" package in the project. To work around the issue,
rename the package.

Signed-off-by: Quentin Monnet <qmo@qmon.net>