refactor: Prefix packages names with "dataplane-"
To avoid any future confusion as faced with the FOSSA tool that picks
dependencies to analyse from crates.io instead of finding the right
packages in the repository, and more generally, for consistency, prefix
the names of all packages from the repo with "dataplane-" in their
respective manifests.

The packages are then added as dependencies without the prefix, meaning
we don't need to change the include directives everywhere in the code,
or the actual paths to the packages source code.

Signed-off-by: Quentin Monnet <qmo@qmon.net>