refactor(nat): Remove (most of) the NAT implementation
Remove all parts of the current implementation that are related to the
representation of user API objects, as well as the parts related to our
own internal IpList structure.

Also remove all API-based definitions from the NAT submodule, as well as
our IpList implementation. A stub IpList submodule is left behind to
avoid breaking compilation, and because we'll provide a new
implementation for that structure in a future commit.

The remaining NAT code is left non-functioning after this commit (but
then, we didn't have fully working NAT so far anyway).

Instead of the deprecated NAT implementation, future commits will update
the submodule to switch to a new implementation based on the actual user
API, and using the table definitions that we've introduced in the latest
commits.

Signed-off-by: Quentin Monnet <qmo@qmon.net>