chore(prefix): Reserve some panicking "From" implementations for tests
The implementations of the From trait for Ipv4Prefix and Ipv6Prefix from
a Prefix may panic if the inner prefix in Prefix is not of the right IP
version. This makes them dangerous to use in sensitive portions of the
code: reserve them for building context for tests, by moving them behind
the "testing" feature.

Signed-off-by: Quentin Monnet <qmo@qmon.net>