ci: Reorder steps in bump.yml workflow to fix upgrades from Git
We have an issue in the bump.yml workflow. It fails to update recursive
dependencies for the project when these dependencies are passed as links
to Git repositories (for example, a branch name and an URL to a GitHub
repository).

I'm not sure what happens exactly, but there seems to be an issue with
the "cargo upgrade" command: it expects some local index to be populated
when running the upgrade for the Git-based dependency, and calls "cargo
update" under the hood with the "--offline" option, but this command
fails.

We've opened a GitHub issue for the upstream "cargo-edit" project
(providing "cargo upgrade"), but in the meantime, let's re-order the
commands we run: if we do the "cargo update" first, this should populate
whatever index "cargo upgrade" seems to rely on when we try to bump the
dependencies in Cargo.toml.

See https://github.com/githedgehog/dataplane/issues/491 for more
context.

Fixes: #491
Signed-off-by: Quentin Monnet <qmo@qmon.net>