feat(mgmt): derive MultiIndexMap on VrfConfig
This adds machinery to enforce uniqueness constraints on the fields of the VrfConfigTable (actual enforcement is in the following commit).

The VPC manager cannot tolerate deviation from the uniqueness constraints and also fundamentally lacks the machinery to signal errors back to users regarding uniqueness constraint violations.
As a result, we need to push the error handling closer and closer to the actual user interaction.
This is the first step on that path.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>