fix(mgmt): enforce constraints on VrfConfigTable
VrfConfigTable was not enforcing uniqueness constraints on its components.

This is fixable by changing the VrfConfigTable to a MultiIndexMap which enforces these constraints automatically.

This is necessary because the interface manager enforces these constraints.  Failure to enforce constraints here will cause errors to appear far from their actual cause site as well as inhibiting the ability to return an error to the user.

In the future we should return a `Result` type from `add_vrf_config`.  That said, I'm trying ot keep this PR as small as possible.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>