feat(mgmt): split apply config in 2 sub-functions
The current invocation of vpc manager is infallible, except if
reconciliation takes an unexpected number of iterations, in which
case it panics. Change this to a failure (given that the code will
attempt to roll-back to the previous config) and split the logic in
two functions for clarity. In the future, this may be exploited to
partially roll-back. E.g. if what failed was the reconciliation,
and the frr config was left untouched, the config processor could
just roll-back the vpc-manager changes.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>