fix(mgmt): Use VTEP config instead of guessing
Before we were deriving the VTEP config from loopback info.  But we have the actual config parameters, so we may as well use those downstream.

This commit basically punts the problem up one layer so we can solve it in the next pass of commits.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>

fixup! for the next one

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>
Signed-off-by: Daniel Noland <daniel@githedgehog.com>