feat(mgmt): lock the cpi/router while applying cfg
Lock the CPI while we apply a configuration. Unlocking is automatic
once the guard given by the lock() method goes out of scope.
When locked, no messages from the control-plane will be processed
and, hence, the routing database be updated. This is needed to
avoid processing messages while kernel interfaces change because
the soft state in the dataplane needed to properly process them
may not yet exist.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>