chore(mgmt): Use Ipv4Addr for vtep.local instead of IpAddr
In theory, vtep addresses can be IPv6, in practice, FRR does
not yet support this.

This commit is a tiny step toward a more unified approach to
dealing with value checking and types.  Ideally, we should have
types that represent checked values so that we don't have to
keep checking them at each layer.  So, here, instead of parsing,
checking for IPv4 and a /32 for vtep.local and then just creating
an IpAddr where later we'll have to check for v4 again it is
better to encode that we've checked the IP address by using
Ipv4Addr instead of IpAddr.

Eventually we should move as many checks as early as possible
and construct appropriate types where it isn't too annoying.
However, that is a bigger lift and lower priority so I only do
vtep.local here since I was manipulating that code anyway.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>