feat(routing): add configuration submodule
This submodule is responsible for configuring the vrfs and inter-
faces of the router. To configure a router object, we need to
provide VRF and interface configurations.

This patch contains utilities to build a configuration and the
methods to apply it. A configuration is initially validated, then
applied and, on success, verified. The verification step double-
checks that the internal state after a suposedly successful
enforcement of a config, according to the internal APIs, conforms
to the configuration.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>