feat(net): Add IP-independent .next_header() method to struct Net
Add a .next_header() method to struct Net, in order to return the
protocol/next header number for the packet, independently of the IP
version of the packet.

To implement this method, we need to convert ip.protocol() into a
NextHeader object. To that end, implement From<IpNumber> for NextHeader.

Derive the Debug trait for struct NextHeader, to make it possible to
embed it in structs that derive the trait, in future commits.

Signed-off-by: Quentin Monnet <qmo@qmon.net>