fix(ci): Update compile-env container name after dpdk-sys changes
With commit 879e0e94d629 ("bump(dpdk-sys): automated bump of dpdk-sys")
we recently bumped the commit SHA for the compile environment. This
broke the setup! This is because we had recent changes in the dpdk-sys
repo, that caused the name of the container to change [0], and the
refresh command to fail:

    $ just create-compile-env

    /!\ Latest compile-env not found, try "just refresh-compile-env"

    Unable to find image 'ghcr.io/githedgehog/dpdk-sys/compile-env:d7a124dccb2e0c3d7931b64584f1f06d5d8f321e' locally

We missed it in CI [1], because we recently reshuffled the CI workflows
in the dataplane repository, and apparently we forgot to mark the
workflows as required to merge Pull Requests (this should now be fixed).

Anyway, we also need to fix the container name for the environment
image. Let's do it by appending the profile name, just like dpdk-sys
does now. And do the same for libc-env container names in the justfile
and in test-runner.sh, as well.

[0] githedgehog/dpdk-sys@a5c9ebd9a8f27b48e74b1c7a97aa08b12080eb4d
[1] https://github.com/githedgehog/dataplane/pull/620#issuecomment-2995983655

Signed-off-by: Quentin Monnet <qmo@qmon.net>