fix(ci): Make Summary step functional again
In a recent commit, in an attempt to "de-clutter" the dev.yml workflow,
we broke the Summary job: it doesn't run all the time, and always
returns 0 anyway. Let's fix it: make it dependent on the results from
the main job, so that we get meaningful results even if we skip most of
the CI jobs when updating the README.md, for example.

This should save us from missing test failures before merging Pull
Requests, as recently happened with a compile-env update.

Fixes: a7b02abbb309 ("test(ci): de-clutter dev.yml workflow")
Signed-off-by: Quentin Monnet <qmo@qmon.net>