refactor(nat): Dissociate stateless NAT from stateful NAT
Most of the code for stateful and stateless NAT is already dissociated,
but the two modes share a common infrastructure whitin the "Nat" object.
This commit separates them entirely, by duplicating the shared code to
create a copy in the "stateless" submodule, and introducing a new
StatelessNat object. The remaining code in lib.rs for stateful NAT is to
be moved in a future commit.

Motivations for separating the two modes into distinct pipeline stages
include:

- They have little code in common anyway
- It will be clearer, when programming the dataplane, to explicitly
  handle a stateless or statefull object
- The structures used by the two modes, in terms of configuration and
  internal state (for stateful NAT), are not the same. As a consequence,
  having both modes in a same object means instances need to keep some
  useless empty context for the chosen running mode, and it also makes
  it more difficult to protect the different objects for concurrent
  access.

Signed-off-by: Quentin Monnet <qmo@qmon.net>