style(nat): Fix up Clippy directives and reports
Following the move of the NAT code to a dedicated crate, let's make sure
we pass the correct directives to Clippy and for rustdoc. Also fix the
minor reports for Clippy, and make it ignore "dead_code" only in the
stateful module where it's relevant, not the whole crate.

Also fix two rustdoc warnings emitted from the "routing" package.

Signed-off-by: Quentin Monnet <qmo@qmon.net>