feat: introduce tc chain
A chain is basically a list of filters sorted in descending order of priority.
A network device (more specifically, a qdisc on a network device) or block may have multiple chains attached to it.
Packets start processing on the qdisc's or block's chain 0, and may or may not move to another chain (depending on which actions they encounter).
jump and goto chain are among the possible actions.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>