feat: introduce tc block
Support the manipulation of tc blocks.

A block is basically a list of qdiscs on different network interfaces.

A filter may be attached to either a qdisc or a block of qdiscs.
This is useful in our case!

Imagine you have 10 network interfaces which are all part of the same underlying physical switch ASIC.
If you configure tc filters in the most basic way, you may need to install the same rule to the ASIC 10 times.

This is bad for a few major reasons:

1. It is pointlessly complex.  You would need to keep track of all 10 rules.  You need to create it 10 times, delete it 10 times, update it 10 times, track the counters 10 times, and so on.
2. It is wasteful of resources in the ASIC.  Installing the rule into the hardware once saves TCAM or SRAM in the ASIC (both of which are in limited supply).
3. It is more prone to race conditions.  Updating the rule in one place is usually much easier to sequence when transitioning between rule sets.

Using blocks allows us to side-step all of these issues.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>