feat: introduce tc qdisc
A qdisc is short for queuing discipline

There are many different algorithms for queueing packets, but the non-trivial case pretty much exclusively focuses on the order in which packets egress a network device.

qdisc also applies to ingress traffic, but in a much more trivial sense.
You can't really control the order or timing in which packets arrive, so the ingress qdisc is mostly just a place to attach filters, chains, and blocks.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>