chore: split nat to avoid circ. dependency w/ mgmt
We need to avoid circular dependencies between nat and mgmt crates.
Do so by moving the definitions related to NAT to the nat crate and
let the mgmt crate consume those. This is a horrible commit due to
the difficulty of making this separation. There exists valid logic
that uses definitions from both mgmt and nat. Such code cannot be
moved to the nat crate and remains in mgmt. Also, there is a test
the exercises stateless nat logic that remains in the mgmt crate.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>