feat(nat): Make NAT stage direction-independant
Remove the enum NatDirection, and all related logics. Instead, have the
stateless NAT pipeline stage do both source and destination NAT, based
on the contents of the NAT tables.

The NAT tables need to be updated accordingly, in a future commit.

Translation in stateful NAT has not been adjusted yet, this will also
come in a future commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>