test(nat): Fix stateless NAT test for forward and reverse paths
Rather than having two different tests for source and destination NAT,
make sure that we try both the reply and request packet, thus validating
NAT in both directions.

Note that the test is not working at this stage: our objective is to
make it work by fixing the code. We temporarily compile out the "tests"
module to avoid failing tests.

Signed-off-by: Quentin Monnet <qmo@qmon.net>