fix(nat): Fix "remote VPC" lookup for stateless source NAT
In the previous commit, we fixed one bug for the source NAT lookups: we
made the NAT table use the "as_range" prefixes rather than the "ips"
prefixes for the virtual "remote peer VPC" lookup, in order to match the
destination address of packets against public IPs for the remote VPC.

This was not enough to fix NAT!

We have two other issues:

- First, the key we use for the "peering_index" in the PrefixTrie lookup
  are the "as_range" of the _local manifest_, which is wrong. We should
  have the prefixes for the "as_range" of the exposes of the _remote_
  manifest, so we need to move the "peering_index" insertion (when
  building the table) to the loop where we process the _remote_ exposes.
- This leads to the second issue: we don't have a 1:1 mapping between
  the VpcExpose objects of the two manifests, which means that a
  destination address corresponding to a given remote peer VPC can
  correspond to any of the configured VpcExpose objects for the peering
  involving this remote VPC. So we can't retrieve a single index, but a
  list of indices.

Let's fix the code accordingly. Move the insertion of peer indices to
the section in "add_peering()" where we process the remote "expose"
blocks. There, make sure we insert a vector containing indices for all
"expose" objects of the local VPC for the peering. Adjust the insertion
and lookup logics accordingly in the "nat" crate, too.

Signed-off-by: Quentin Monnet <qmo@qmon.net>