fix(nat): Use ips if as_range are empty for stateless src NAT 1st lookup
We have yet another issue when building the tables: if we do NAT on one
side only, we won't populate the data necessary for doing source NAT for
the remote VPC, because we don't find any as_range to use. What we need
to do is to fall back on "ips" when there's no as_range prefixes,
because this is where the public prefixes are contained in that case.

We also have an issue where we don't find a mapping in the list for
destination NAT, which is easily solved by not creating destination NAT
rules at all if we don't have any as_range for a given expose.

Reported-by: Manish Vachharajani <manish@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>