feat(nat): Collapse prefixes when building stateless NAT table
Rather than optimizing the lists for prefixes by only removing exclusion
prefixes when they cover one half of allowed prefixes, split allowed
prefixes into smaller fragments to get rid of all exclusion prefixes
altogether.

We don't spend too much effort on the new error introduced in this
commit; it's more convenient at the moment to return a TrieError from
the function (because we return it in several locations), so just add an
"Other" error for the case where the internal algorithm has a bug.

Signed-off-by: Quentin Monnet <qmo@qmon.net>