feat: introduce tc actions
An action is some type of manipulation or event that may occur when a packet matches on a [filter] criteria.

Examples of action include

* dropping packets,
* editing source or destination ip addresses,
* pushing or popping VLAN headers,
* redirecting the packet to the ingress or egress pipeline of another network device,
* encapsulating the packet in a VXLAN packet,
* or mirroring (copying) the packet to another network device.

Actions are reference counted and may be attached to more than one filter.
More specifically, each action has four important variables associated with it

1. The action `kind`.  `kind` is a static string which identifies the family of the action.  Examples include `mirred`, which covers packet redirect and mirroring; `gact`, which includes actions like `drop` or `jump`; and `tunnel_key`, which includes encapsulation and decapsulation actions.  Many other families of actions exist in the linux `tc` subsystem.
2. The action `index`.  `index`, is a unique (_per `kind`_) identifier to track the specific incidence of a given action.  For example, if you might create a `gact` dro action with index 17.  You may then attach that drop command to several different filters by referencing that index on filter creation.  If you recycle an action in this way then all the filters associated with that action will update the hit counters for that action.  Reuse of actions in this way may also save resources in the network card (assuming the matches and action can be offloaded).
3. The action `bindcnt`.  This is the number of active filters which reference this action.
4. The action `refcnt`.  This is the reference count of the action.  This number is either equal to the `bindcnt` or is equal to the `bindcnt` plus one.

If you create an action as part of a filter creation command, then you the `refcnt` and the `bindcnt` numbers will be equal.
If you create an action without attaching it to a filter, you need to set the `refcnt` to one (in which case the `bindcnt` will be zero automatically).

Linux will automatically remove any action if and only if the `bindcnt` _and_ `refcnt` are equal to zero.
"Deleting" an action sets its `refcnt` to its `bindcnt`.
Note that this means you need to both delete an action _and_ detach that action from all active filters before linux will actually remove the action.
This is a good thing!
It prevents you from kicking the legs out from under your own network card by deleting an action (some actions make no sense without their accompanying actions).

Signed-off-by: Daniel Noland <daniel@githedgehog.com>