feat(nat): Add proper error handling for IpList (stateless NAT)
We have a few calls to unreachable!() in the IpList logic, where we
compute and map the offsets from and to IP addresses within lists of
prefixes for stateless NAT.

Those unreachable!() have been reached in the past.

We also have assumptions that the offsets or IP addresses that we pass
to the functions are always contained within the lists or prefixes that
we process, which is only true as far as we don't have any bug in the
logic.

Instead of these, let's introduce proper error handling for the module.
When something's wrong, we raise it to the caller. This should avoid
crashing the whole dataplane (you never know) and should help debug in
case of bugs.

The errors propagate to the NAT translation functions, where we stop
translating the address when we fail to compute the mapping. We need to
improve error handling in there as well, and to drop the packet when we
face issues with the logic of the processing during NAT, but this is not
covered by this commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>