feat(mgmt): Add helper function to handle publicly-exposed IPs for NAT
A given VpcExpose may or may not required NAT, depending on whether the
as_range list contains any prefix or not. This has consequences in some
portions of the NAT processing logics: at times, we need to consider the
publicly-exposed addresses for an expose object, and these may not be in
as_range, but in ips, if NAT is not in use for this expose.

Add some helper methods to handle public IPs in a more convenient way.
This makes the logics easier to follow when we need to refer to these
public IPs whether they are in as_range or ips.

Note that this commit does not account for the case where we have an
empty as_range list but non-empty not_as, given that this case is
currently rejected at validation time.

Suggested-by: Manish Vachharajani <manish@githedgehog.com>
Signed-off-by: Quentin Monnet <qmo@qmon.net>