chore(nat): Remove unused iplist submodule
The iplist module has been rewritten, it disappeared at some point, then
it came back... but this time, it looks like we no longer need it, and
we remove it again. For good? Time will tell...

This is a follow-up to a rework of the stateless NAT tables, and the
module is now no longer in use.

Signed-off-by: Quentin Monnet <qmo@qmon.net>