feat(mgmt): Simplify and optimise TrieValue generation
As part of the code for adding the contents of a Peering to a NAT table,
we process the different prefix lists of the VpcExpose objects to add
them to the relevant tables.

The way we've been doing that is by looping over each prefix, and
calling get_public_trie_value() or get_private_trie_value() for that
prefix; and then in these functions, we'd loop again over all prefixes
to find the offset in the list of the prefix we're processing. This is
O(n^2) where n is the number of prefixes in the VpcExpose list. There's
no need for the double loop, we can just as well iterate once and
compute the offsets as we go.

Change the TrieValue generation code to do this. We make the functions
return iterators, so that the caller can just iterate on tuples with
(prefix_key, trie_value), and add them to the map. We also move the size
checks to a dedicate function. These changes make the code easier to
follow, and get us read of the double loops.

Signed-off-by: Quentin Monnet <qmo@qmon.net>