refactor(mgmt): Rename TrieValue to NatTableValue
The NatTableValue is no longer associated with a trie (we use a
BTreeHash and retrieve the lower bound for an IP to find the IP range
start address). Now, the name is misleading: let's rename the struct
(and some associated functions: generate_trie_values(),
generate_public_trie_values(), generate_private_trie_values()).

Signed-off-by: Quentin Monnet <qmo@qmon.net>