feat(prefix): Implement SubAssign trait for PrefixSize, fix Sub impl
Implement trait SubAssign for PrefixSize, given that we already suppot
the Sub trait.

Speaking of, we also adjust our implementation of the Sub trait for
PrefixSize - PrefixSize, so that subtracting Overflow from something
triggers an overflow panic (instead of returning Overflow). Also make
sure that we panic and never wrap in case of subtaction overflow (it
doesn't make sense to wrap around in case of PrefixSize objects).

Signed-off-by: Quentin Monnet <qmo@qmon.net>