feat(config): add new crate 'config'
The new crate 'config' is meant to contain the dataplane configu-
ration model and types outside of the mgmt crate so that other
crates can refer to those types. The mgmt crate will, nevertheless,
continue to be the only entity responsible for building, storing
or distributing configurations to other functional components
within the dataplane.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>