feat(mgmt): remove build_nat_internal_config()
Remove that function from the mgmt crate and use instead the
function in the nat crate. This is fine because such a function
is unlike the other functions in mgmt very little opinionated and
just iterates over vpcs and calls add_peering(), which no longer
needs to be public. With this, the mgmt crate has minimal
dependencies on the nat crate.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>