feat(routing): add event log infrastructure
Defines an EventLog type and several macros to instantiate
EventLogs as thread-local. An EventLog is just a generic collection
of ordered, timestamped events. The implementation is generic in
that the type for the events is not prescribed, so that multiple
types can be used with distinct EventLogs in the same or different
threads.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>