refactor(nat): Remove now unused PrefixTrie module
In commit 7915a4c30822 ("feat(nat): Use dst_vni metadata to simplify
stateless NAT"), we removed the last portions of code that would use the
struct PrefixTrie in the nat crate. The prefixtrie submodule is now
unused, let's remove it. If we need a trie again in the future, we can
now use our lpm package anyway.

Signed-off-by: Quentin Monnet <qmo@qmon.net>