feat(routing): fibgroup removal from map
Fibgroup removal is tricky because fibgroups are shared by
fibroutes and mutable. We need to ensure that we don't remove
a fibgroup from the store if it is referenced by a fibroute.
If we did, the route would still point to the fibgroup, but
we would not be able to mutate it; hence the route would not
be updated if the fibgroup changed. On the other hand, we don't
want to leave stale fibgroups around.

Ideally, we would delete a fibgroup once the corresponding next-hop
was deleted. This patch adds a fib method to unregister a group,
called whenever a next-hop is destroyed. However, fibgroups are
not deleted if still have references to them. Therefore, as
a sanity, the purge method is still called. The reason for this is
to make sure we never delete a fibgroup that is referenced. This
should be easy to check, but there's a problem: because left-right
clones the underlaying data structure, refcounts get "doubled".
While this could be compensated changing the removal criteria
(remove if refcount reaches 2 instead of 1), for the time being,
the safest is to use the explicit unregister method introduced
in this patch, along with a call to purge() after fib route
deletions. By keeping the original removal criteria (which is
more restrictive) we ensure that we don't delete fibgroups
prematurely. By purging, we ensure that we will end-up deleting
those that are no longer in use, though not immediately.

Probably, the best approach may be to mark fibgroups as deletable
with the explicit unregister call when the corresponding next-hop
is removed, and delete those that get marked once no route referen-
ces them. However, this requires augmenting augmenting the
fibgroups structure with some flag, and the current implementation
works fine.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>