ci: Remove job to update dpdk-sys (do it from dpdk-sys repo)
We have recently introduced a new workflow in the dpdk-sys repository to
update the package reference in the dataplane repository on new pushes
to the main branch. This makes it easier to get updates as soon as
they're merged in dpdk-sys, without having to run the workflow every day
just to check updates availability.

As a consequence, we no longer need the job in the dataplane repository.
Remove it.

Signed-off-by: Quentin Monnet <qmo@qmon.net>