fix(vpc-manager): Avoid hitting unreachable!
The dataplane may receive configurations that do not include any
VPC. In that case (i.e. underlay-only configs) we can't assume
that the vtep configuration is there as it is not needed.
So, the vpc manager should only assume that a vtep config is
present if there are VPCs (vrfs aside from the default).

The original code worked fine because the default vrf was
excepted by the vpc manager: if there were no vrfs other than
the default, we were not checking for the existence of a vtep,
and if there were, the validation of the configuration ensured
that a vtep was present.

A guard for this was removed in
3301da3ec51f71ceb98339dea97d2d4237ab8c6b
when the vpc manager started to create interfaces for the default
vrf, which broke the original assumption. Restoring the guard that
excepted the default vrf is no longer an option since we need to
configure interfaces for the default vrf. So, split the method to
iterate over non-default vrfs and then deal with the interfaces
for the default vrf.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>