chore(nat): Use VNI instead of VRF ID as VPC identifier in stateful NAT
We may reconsider this in the future (Daniel mentioned RouteTableId as a
VPC discriminant), but for now we use VNIs to identify the VPCs we're
dealing with in the stateful NAT code. Make the code easy to adjust if
we use something else in the future, though.

We need to pass identifiers for both the source and destination VPC
(hence two VNIs) to perform NAT correctly: adjust accordingly.

Signed-off-by: Quentin Monnet <qmo@qmon.net>