chore(nat): Add L4 ports to stateful NAT tuples
We need to take L4 ports into account as part of the tuple that we
extract for a given packet to correctly look up the associated flow in
the session map. Add the port wherever needed.

Note: L4 support is not complete yet at this stage (see FIXME in the
commit), we're missing L4 ports retrieval for the packet. This will be
another time.

Link: https://github.com/githedgehog/dataplane/issues/800
Signed-off-by: Quentin Monnet <qmo@qmon.net>