feat(nat): Add IP allocation to stateful NAT allocator
This is the stage where we add IP allocation to our new allocator object
for stateful NAT.

We get the IpAllocator, which contains a NatPool with PoolBitmap objects
to mark Ip addresses as available or not, and a list of weak references
to AllocatedIp objects, representing allocated addresses. An AllocatedIp
contains a PortAllocator object to further allocate ports, but this
component is not implemented in this commit.

Internally, IPs are represented by a large (but compressed) bitmap,
containing u32 integers. This is handy for IPv4, given that a bitmap
just covers the IP space, but not as much for IPv6. We deal with this
situation by associating offets in the bitmap to offsets (addresses) in
the list of IPv6 addresses in use for NAT. Of course, we can't repreent
all possible IPv6 addresses from the bitmap, so we have more than 2^32
addresses in the prefix lists - say if the user passed a /64 as a prefix
for NAT - then we just ignore the remaining ones. In practice, given the
amount of memory we need to store for each allocated IP, we'll hit
hardware limitations way before we allocate 4 billion IP addresses, so
the workaround should work just fine.

Working with a bitmap in generic functions calls for methods specific to
IPv4 and IPv6 when it comes to mapping the addresses to bitmap values.
To deal with this, we introduce a new trait, NatIpWithBitmap, that
overloads NatIp.

Signed-off-by: Quentin Monnet <qmo@qmon.net>