feat(nat): Add IPv6 support to stateful NAT allocator
Given that the codebase for stateful NAT has been moving quite a lot
during the developement phase, we focused on IPv4, but haven't
implemented the IPv6 counterpart functions. Do it now. It's easy: it's
the same as for IPv4, transposed with IPv6-specific tables or functions
at a couple locations.

Signed-off-by: Quentin Monnet <qmo@qmon.net>