test(nat): Introduce shuttle compatibility for stateful NAT
Make the default allocator used components from our "concurrency" crate
for concurrency-related objects, in order to make the code compatible
with shuttle-based tests, shuttle being a framework to test code with
concurrent access under random scheduling conditions.

Add two unit tests: One is a simple test to make sure that basic shuttle
tests work (it does not contain NAT-related elelments); the second uses
std or shuttle components (depending on how it is compiled) and serves
as a "control" test to validate whether a failure is related to shuttle,
or not (for example: it allowed us to detect that the repeated
stackoverflow at runtime was not our fault - the snippet worked with std
components - but the one from the library, with a default of 32 kB stack
quickly overflowed due to the large size (100+ bytes) of shuttle's
AtomicBool objects.

Note: The simple shuttle test is not called in CI, yet.

Signed-off-by: Quentin Monnet <qmo@qmon.net>