chore: {src,dst}_vni to {src,dst}_vpcd
Note that this does not enable use of non-VNI VpcDiscriminants.
All it does is change the interfaces to extract the VNI from
the VpcDiscriminat and then use it, or error otherwise, as if
there were no Discriminant.  THis is fine for now as VNI is
the only supported discriminant type, but will need a revisit
if we ever support MPLS or something like that.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>