test(nat): Remove useless "tests" module in test_alloc.rs
The whole file contains tests, now that "standard" (non-shuttle) tests
are inside of their own "std_tests" submodule there's no good reason to
keep the outer "tests" module. Let's remove the "tests" module.

Most of this commit is just indentation changes as a result of the
removal of the outer "tests" module - on GitHub, this is better reviewed
by hiding whitespace changes.

Signed-off-by: Quentin Monnet <qmo@qmon.net>