test(shuttle): Run shuttle tests in CI
Run tests using Shuttle in CI.

These tests require recompiling the code with the "shuttle" feature, so
we cannot tweak the "cargo nextest" invocation to run them at the same
time as the regular tests, we run a second "cargo nextest" invocation.
Happily, the results.json produced when running the tests does not
contains a full JSON object, but a list of JSON objects, so we can
safely append to the file and still look for flakes without changing the
existing workflow code that parses the file to look for flakes.

To avoid running all tests a second time, we filter tests to run using
the "shuttle" pattern. This means that all tests expected to run with
shuttle need to have the "shuttle" pattern in their test name - this can
be in their module name, their submodule name, or the name of the test
function.

Signed-off-by: Quentin Monnet <qmo@qmon.net>