chore(nat): Expose default allocator publicly
If we want to build and reference this allocator from a crate other than
NAT, we need to expose the objects publicly.

Fix the rustdoc errors that come up when we make the module public. We
allow referencing private objects in the docs, as the internals for the
default allocator aren't really expected to be used publicly other than
when building the allocator.

Signed-off-by: Quentin Monnet <qmo@qmon.net>