feat(mgmt,nat): Add writer for stateful NAT allocator
In preparation for building the stateful NAT allocator, add a writer to
the ConfigProcessor and the InternalSetup objects. This writer mimics
what we do for stateless NAT, although its implementation will differ,
because we're not using left-right for stateful NAT (too many writes to
update the allocator's internal data as we track connections).

Note that the writer is currently not generic (with regards to the type
of the allocator), as this would require propagating generic arguments
and exposing the NatDefaultAllocator in many locations, for little
benefit. If we want to swap for another allocator, it should remain easy
enough to update the code in allocator_writer.rs.

Signed-off-by: Quentin Monnet <qmo@qmon.net>