ci: Fail on rustdoc warnings
There seems to be no option to make docs generation fail on warnings,
and we've missed a few in CI in the past - latest example to date was
fixed in commit 07ca7b537470 ("fix(pkt-meta): Tame rustdoc warnings").

There is a way to make the build fail on warnings, though: we need to
set the RUSTDOCFLAGS appropriately. Let's do it to have CI raise
warnings in the future.

Signed-off-by: Quentin Monnet <qmo@qmon.net>