fix(nat): Fix port selection when creating reverse session entry
When we create the return session for a connection, we pay attention at
the IP addresses to use: if no NAT is required for one of the two
addresses (source/destination), then we use the original address to
create the return session. This is what happens, for example, in the
following mapping:

    forward: [1.1.2.3:9998     -> 10.201.201.18:443] NATs to [10.12.0.0:39680   -> 1.2.2.0:51200]
    return:  [1.2.2.0:51200    -> 10.12.0.0:39680]   NATs to [10.201.201.18:443 -> 1.1.2.3:9998]

The return session entry is a mirror from the forward path entry, this
is good.

Now, if we do not apply NAT in one of the two directions, say for the
destination address of the request, we get the following session entries
created:

    forward: [1.1.2.3:9998     -> 10.201.201.17:80]  NATS to [2.2.0.0:40704     -> 0.0.0.0:0]
    return:  [10.201.201.17:0  -> 2.2.0.0:40704]     NATS to [0.0.0.0:0         -> 1.1.2.3:9998]

The source address to NAT on the return path is correct. It corresponds
to the initial destination address, given that it is never NAT-ed.
However, note that the source port in the entry is set to 0 instead of
port number 80. This is a mistake: we used the target destination port
from the forward path (0, given that there's no NAT required), and the
entry for the return path will never actually match the replies. The
reply coming back fails to match an existing session entry, and
generates a new entry after that:

    forward: [10.201.201.17:80 -> 2.2.0.0:40704]     NATS to [0.0.0.0:0         -> 1.1.0.0:30976]
    return:  [1.1.0.0:30976    -> 10.201.201.17:0]   NATS to [2.2.0.0:40704     -> 0.0.0.0:0]

To fix this issue, we need to process ports just like IP addresses: use
the original ones if there's no NAT to do for either the source or
destination address and port. We already have the logics in place for IP
addresses, just integrate the ports.

Note: This code will likely become legacy soon, given that the new flow
table implementation, not used for NAT yet, provides bidirectional
sessions creation.

Fixes: 8b2267c7df60 ("feat(nat): Make stateful NAT pipeline stage use the default allocator")
Signed-off-by: Quentin Monnet <qmo@qmon.net>