feat(nat): Implement Display trait for NAT sessions
The current code for NAT sessions will likely be dropped soon, as we
intend to move the implementation to the newer flow table implementation
we have in the code base.

Nevertheless, I needed to implement the Display trait to facilitate
debugging of a test, and I don't like to waste it. Add it to the session
code.

Signed-off-by: Quentin Monnet <qmo@qmon.net>