refactor(dataplane): Merge start_router() and setup_routing_pipeline()
The building of the pipeline was done in two steps: start_router() would
initialise all necessary objects, and then call setup_routing_pipeline()
to build the network functions, and at last, the pipeline itself.

This separation makes sense, but is not convenient: At the moment, we
hard-code the pipeline stages that we want in the pipeline, which means
that we may, or may not require some of the objects initialised in
start_router(). As a consequence, each time we modify the pipeline, we
may have to change the interface and the call to
setup_routing_pipeline(), to adjust the list of arguments accordingly.

Instead of doing so, merge the two functions. The resulting
start_router() function remains clear enough, and this way we can only
instanciate the objects we need, without having to update the parameters
of an intermediate function.

Signed-off-by: Quentin Monnet <qmo@qmon.net>